+++ src/webgis4u/ol/util.js
@@ -2,6 +2,7 @@
  * @module webgis4u/ol/util
  */
 
+export * from './util/createImageFromMap';
 export * from './util/getMapType';
 export * from './util/getScale';
 export * from './util/pan';