+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
   * [`number`](./src/webgis4u/util/number.js):
     * [`convertUnitValue`](./src/webgis4u/util/number/convertUnitValue.js):
   * [`ol`](./src/webgis4u/util/ol.js):
+    * [`createImageFromMap`](./src/webgis4u/ol/util/createImageFromMap.js)
     * [`getMapChildElementByClassName`](./src/webgis4u/util/ol/getMapChildElementByClassName.js):
     * [`getMapType`](./src/webgis4u/ol/util/getMapType.js)
     * [`getScale`](./src/webgis4u/ol/util/getScale.js)